♻️ chore: extract KubernetesResource schema for future broader usage
