+++ lib/modules/manager/flux/schema.ts
@@ -1,16 +1,5 @@
 import { z } from 'zod';
-
-export const KubernetesResource = z.object({
-  apiVersion: z.string(),
-  kind: z.string(),
-  metadata: z.object({
-    name: z.string(),
-    // For Flux, the namespace property is optional, but matching HelmReleases to HelmRepositories would be
-    // much more difficult without it (we'd have to examine the parent Kustomizations to discover the value),
-    // so we require it for renovation.
-    namespace: z.string().optional(),
-  }),
-});
+import { KubernetesResource } from '../kubernetes/schema';
 
 export const HelmRelease = KubernetesResource.extend({
   apiVersion: z.string().startsWith('helm.toolkit.fluxcd.io/'),
