+++ lib/modules/manager/kubernetes/schema.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod';
+
+export const KubernetesResource = z.object({
+  apiVersion: z.string(),
+  kind: z.string(),
+  metadata: z.object({
+    name: z.string(),
+    namespace: z.string().optional(),
+  }),
+});
